requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
pyrogram==2.0.106
TgCrypto==1.2.5          # C-extension for Pyrogram (faster crypto)

# ── Database ──────────────────────────────────────────────────────────────────
motor==3.3.2              # Async MongoDB driver
pymongo==4.6.1            # Motor dependency

# ── Image processing ─────────────────────────────────────────────────────────
Pillow==10.3.0

# ── HTTP ─────────────────────────────────────────────────────────────────────
aiohttp==3.9.5
aiofiles==23.2.1

# ── Config ───────────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Utility ──────────────────────────────────────────────────────────────────
numpy==1.26.4             # Used in gradient generation (optional but fast)
